# application/services/custom_services/reconcile/src/models/entitymodels/discrepancy.ts
export interface discrepancy 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   discid: { type: String },
   message: { type: String },
   type: { type: String }
}